Refactor, focused on naming
