+++ producer/kafka_producer.go
@@ -1,4 +1,4 @@
-package producers
+package producer
 
 import (
 	"log"
